Added check when evencode buffers overlap
When the `startEventCodeBufferMS` overlaps in time with the
`endEventCodeBufferMS`, the overlapping time window is erroneously
marked for rejection, even though that overlapping time window is
supposed to be protected.

For example, if the user specifies:

2000 ms threshold
1500 ms startEventCodeBufferMS
2000 ms endEventCodeBufferMS

and the function finds a 2100 ms time segment to delete. Because
`startEventCodeBufferMS` and `endEventCodeBufferMS` overlap, there
should be no time segment to delete. However, as previously constructed
the function deletes the overlapping time segment.

This new function ensures that the time segment is preserved.